When sending out a heartbeat, also send a ping frame
My current implementation of the client can be quiet for hours, which would
trigger the 60-second timeout in `cowboy_websocket`. Adding a ping-frame
forces clients to respond with a pong-frame, restarting the loop. In testing
I found that some browsers work better with some payload data added, so I'm
adding the same "h" string as payload.